+++ packages/0x.js/CHANGELOG.md
@@ -8,6 +8,8 @@ vx.x.x
     * Remove `ZeroExError.ContractNotFound` and replace it with contract-specific errors (#233)
     * Make `DecodedLogEvent<A>` contain `LogWithDecodedArgs<A>` under log key instead of merging it in like web3 does (#234)
     * Rename `removed` to `isRemoved` in `DecodedLogEvent<A>` (#234)
+    * Add config allowing to specify gasPrice and gasLimit for every transaction sending method (#235)
+    * All transaction sending methods now call `estimateGas` if no gas amount was supplied (#235)
     * Modify order validation methods to validate against the `latest` block, not against the `pending` block (#236)
 
 v0.26.0